feat: fif integration tests split into files
