+++ test/spec/FindInFiles-test.js
@@ -175,7 +175,7 @@ define(function (require, exports, module) {
             runs(function () {
                 var $searchField = $("#find-what");
                 $searchField.val(searchString).trigger("input");
-                SpecRunnerUtils.simulateKeyEvent(KeyEvent.DOM_VK_RETURN, "keydown", $searchField[0]);
+                //SpecRunnerUtils.simulateKeyEvent(KeyEvent.DOM_VK_RETURN, "keydown", $searchField[0]);
             });
             waitsFor(function () {
                 return FindInFiles._searchDone;
@@ -442,15 +442,31 @@ define(function (require, exports, module) {
                 var fileEntry = FileSystem.getFileForPath(testPath + "/foo.js");
                 openSearchBar(fileEntry);
                 executeSearch("foo1");
+                runs(function () {
+                    var $searchField = $("#find-what");
+                    SpecRunnerUtils.simulateKeyEvent(KeyEvent.DOM_VK_RETURN, "keydown", $searchField[0]);
+                });
                 executeSearch("foo2");
+                runs(function () {
+                    var $searchField = $("#find-what");
+                    SpecRunnerUtils.simulateKeyEvent(KeyEvent.DOM_VK_RETURN, "keydown", $searchField[0]);
+                });
                 executeSearch("foo3");
+                runs(function () {
+                    var $searchField = $("#find-what");
+                    SpecRunnerUtils.simulateKeyEvent(KeyEvent.DOM_VK_RETURN, "keydown", $searchField[0]);
+                });
                 executeSearch("foo4");
+                runs(function () {
+                    var $searchField = $("#find-what");
+                    SpecRunnerUtils.simulateKeyEvent(KeyEvent.DOM_VK_RETURN, "keydown", $searchField[0]);
+                });
                 executeSearch("foo5");
 
                 runs(function () {
                     var searchHistory = PreferencesManager.getViewState("searchHistory");
-                    expect(searchHistory.length).toBe(5);
-                    expect(searchHistory).toEqual(["foo5", "foo4", "foo3", "foo2", "foo1"]);
+                    expect(searchHistory.length).toBe(4);
+                    expect(searchHistory).toEqual(["foo4", "foo3", "foo2", "foo1"]);
                 });
             });
 
@@ -744,1717 +760,6 @@ define(function (require, exports, module) {
                 runs(function () {
                     // Search model shouldn't be cleared after search bar closed without running a search
                     expect(Object.keys(FindInFiles.searchModel.results).length).not.toBe(0);
-                });
-            });
-        });
-
-        describe("Find results paging", function () {
-            var expectedPages = [
-                {
-                    totalResults: 500,
-                    totalFiles: 2,
-                    overallFirstIndex: 1,
-                    overallLastIndex: 100,
-                    matchRanges: [{file: 0, filename: "manyhits-1.txt", first: 0, firstLine: 1, last: 99, lastLine: 100, pattern: /i'm going to\s+find this\s+now/}],
-                    firstPageEnabled: false,
-                    lastPageEnabled: true,
-                    prevPageEnabled: false,
-                    nextPageEnabled: true
-                },
-                {
-                    totalResults: 500,
-                    totalFiles: 2,
-                    overallFirstIndex: 101,
-                    overallLastIndex: 200,
-                    matchRanges: [{file: 0, filename: "manyhits-1.txt", first: 0, firstLine: 101, last: 99, lastLine: 200, pattern: /i'm going to\s+find this\s+now/}],
-                    firstPageEnabled: true,
-                    lastPageEnabled: true,
-                    prevPageEnabled: true,
-                    nextPageEnabled: true
-                },
-                {
-                    totalResults: 500,
-                    totalFiles: 2,
-                    overallFirstIndex: 201,
-                    overallLastIndex: 300,
-                    matchRanges: [
-                        {file: 0, filename: "manyhits-1.txt", first: 0, firstLine: 201, last: 49, lastLine: 250, pattern: /i'm going to\s+find this\s+now/},
-                        {file: 1, filename: "manyhits-2.txt", first: 0, firstLine: 1, last: 49, lastLine: 50, pattern: /you're going to\s+find this\s+now/}
-                    ],
-                    firstPageEnabled: true,
-                    lastPageEnabled: true,
-                    prevPageEnabled: true,
-                    nextPageEnabled: true
-                },
-                {
-                    totalResults: 500,
-                    totalFiles: 2,
-                    overallFirstIndex: 301,
-                    overallLastIndex: 400,
-                    matchRanges: [{file: 0, filename: "manyhits-2.txt", first: 0, firstLine: 51, last: 99, lastLine: 150, pattern: /you're going to\s+find this\s+now/}],
-                    firstPageEnabled: true,
-                    lastPageEnabled: true,
-                    prevPageEnabled: true,
-                    nextPageEnabled: true
-                },
-                {
-                    totalResults: 500,
-                    totalFiles: 2,
-                    overallFirstIndex: 401,
-                    overallLastIndex: 500,
-                    matchRanges: [{file: 0, filename: "manyhits-2.txt", first: 0, firstLine: 151, last: 99, lastLine: 250, pattern: /you're going to\s+find this\s+now/}],
-                    firstPageEnabled: true,
-                    lastPageEnabled: false,
-                    prevPageEnabled: true,
-                    nextPageEnabled: false
-                }
-            ];
-
-            function expectPageDisplay(options) {
-                // Check the title
-                let match = $("#find-in-files-results .title").text().match("\\b" + options.totalResults + "\\b");
-                expect(match).toBeTruthy();
-                match = $("#find-in-files-results .title").text().match("\\b" + options.totalFiles + "\\b");
-                expect(match).toBeTruthy();
-                var paginationInfo = $("#find-in-files-results .pagination-col").text();
-                match = paginationInfo.match("\\b" + options.overallFirstIndex + "\\b");
-                expect(match).toBeTruthy();
-                match = paginationInfo.match("\\b" + options.overallLastIndex + "\\b");
-                expect(match).toBeTruthy();
-
-                // Check for presence of file and first/last item rows within each file
-                options.matchRanges.forEach(function (range) {
-                    let $fileRow = $("#find-in-files-results tr.file-section[data-file-index='" + range.file + "']");
-                    expect($fileRow.length).toBe(1);
-                    let fileName = $fileRow.find(".dialog-filename").text();
-                    expect(fileName).toEqual(range.filename);
-
-                    let $firstMatchRow = $("#find-in-files-results tr[data-file-index='" + range.file + "'][data-item-index='" + range.first + "']");
-                    expect($firstMatchRow.length).toBe(1);
-                    match = $firstMatchRow.find(".line-number").text().match("\\b" + range.firstLine + "\\b");
-                    expect(match).toBeTruthy();
-                    match = $firstMatchRow.find(".line-text").text().match(range.pattern);
-                    expect(match).toBeTruthy();
-
-                    let $lastMatchRow = $("#find-in-files-results tr[data-file-index='" + range.file + "'][data-item-index='" + range.last + "']");
-                    expect($lastMatchRow.length).toBe(1);
-                    match = $lastMatchRow.find(".line-number").text().match("\\b" + range.lastLine + "\\b");
-                    expect(match).toBeTruthy();
-                    match = $lastMatchRow.find(".line-text").text().match(range.pattern);
-                    expect(match).toBeTruthy();
-                });
-
-                // Check enablement of buttons
-                let disabled = $("#find-in-files-results .first-page").hasClass("disabled");
-                expect(disabled).toBe(!options.firstPageEnabled);
-                disabled = $("#find-in-files-results .last-page").hasClass("disabled");
-                expect(disabled).toBe(!options.lastPageEnabled);
-                disabled = $("#find-in-files-results .prev-page").hasClass("disabled");
-                expect(disabled).toBe(!options.prevPageEnabled);
-                disabled = $("#find-in-files-results .next-page").hasClass("disabled");
-                expect(disabled).toBe(!options.nextPageEnabled);
-            }
-
-            it("should page forward, then jump back to first page, displaying correct contents at each step", function () {
-                openProject(SpecRunnerUtils.getTestPath("/spec/FindReplace-test-files-manyhits"));
-                openSearchBar();
-
-                // This search will find 500 hits in 2 files. Since there are 100 hits per page, there should
-                // be five pages, and the third page should have 50 results from the first file and 50 results
-                // from the second file.
-                executeSearch("find this");
-
-                runs(function () {
-                    $("#find-in-files-results .next-page").click();
-                    waitms(2000);
-                });
-                runs(function () {
-                    expectPageDisplay(expectedPages[1]);
-                });
-                runs(function () {
-                    $("#find-in-files-results .next-page").click();
-                    waitms(2000);
-                });
-                runs(function () {
-                    expectPageDisplay(expectedPages[2]);
-                });
-                runs(function () {
-                    $("#find-in-files-results .first-page").click();
-                    waitms(2000);
-                });
-                runs(function () {
-                    expectPageDisplay(expectedPages[0]);
-                });
-            });
-
-            it("should jump to last page, then page backward, displaying correct contents at each step", function () {
-                openProject(SpecRunnerUtils.getTestPath("/spec/FindReplace-test-files-manyhits"));
-
-                executeSearch("find this");
-
-                runs(function () {
-                    $("#find-in-files-results .last-page").click();
-                    waitms(2000);
-                });
-                runs(function () {
-                    expectPageDisplay(expectedPages[4]);
-                });
-                runs(function () {
-                    $("#find-in-files-results .prev-page").click();
-                    waitms(2000);
-                });
-                runs(function () {
-                    expectPageDisplay(expectedPages[3]);
-                });
-                openProject(defaultSourcePath);
-                waitms(2000);
-            });
-        });
-
-        describe("SearchModel update on change events", function () {
-            var oldResults, gotChange, wasQuickChange;
-
-            function fullTestPath(path) {
-                return testPath + "/" + path;
-            }
-
-            function expectUnchangedExcept(paths) {
-                Object.keys(FindInFiles.searchModel.results).forEach(function (path) {
-                    if (paths.indexOf(path) === -1) {
-                        expect(FindInFiles.searchModel.results[path]).toEqual(oldResults[path]);
-                    }
-                });
-            }
-
-            beforeEach(function () {
-                gotChange = false;
-                oldResults = null;
-                wasQuickChange = false;
-
-                FindInFiles.clearSearch(); // calls FindInFiles.searchModel.clear internally
-                FindInFiles.searchModel.on("change.FindInFilesTest", function (event, quickChange) {
-                    gotChange = true;
-                    wasQuickChange = quickChange;
-                });
-
-                openTestProjectCopy(defaultSourcePath);
-                doSearch({
-                    queryInfo: {query: "foo"},
-                    numMatches: 14
-                });
-                waitsFor(function () {
-                    return FindInFiles.searchModel && FindInFiles.searchModel.results;
-                }, 1000, "search completed");
-                runs(function () {
-                    oldResults = _.cloneDeep(FindInFiles.searchModel.results);
-                });
-            });
-
-            afterEach(function () {
-                FindInFiles.searchModel.off(".FindInFilesTest");
-                waitsForDone(CommandManager.execute(Commands.FILE_CLOSE_ALL, { _forceClose: true }), "close all files");
-            });
-
-            describe("when filename changes", function () {
-                it("should handle a filename change", function () {
-                    runs(function () {
-                        FindInFiles._fileNameChangeHandler(null, fullTestPath("foo.html"), fullTestPath("newfoo.html"));
-                    });
-                    waitsFor(function () { return gotChange; }, "model change event");
-                    runs(function () {
-                        expectUnchangedExcept([fullTestPath("foo.html"), fullTestPath("newfoo.html")]);
-                        expect(FindInFiles.searchModel.results[fullTestPath("foo.html")]).toBeUndefined();
-                        expect(FindInFiles.searchModel.results[fullTestPath("newfoo.html")]).toEqual(oldResults[fullTestPath("foo.html")]);
-                        expect(FindInFiles.searchModel.countFilesMatches()).toEqual({files: 3, matches: 14});
-                        expect(wasQuickChange).toBeFalsy();
-                    });
-                });
-
-                it("should handle a folder change", function () {
-                    runs(function () {
-                        FindInFiles._fileNameChangeHandler(null, fullTestPath("css"), fullTestPath("newcss"));
-                    });
-                    waitsFor(function () { return gotChange; }, "model change event");
-                    runs(function () {
-                        expectUnchangedExcept([fullTestPath("css/foo.css"), fullTestPath("newcss/foo.css")]);
-                        expect(FindInFiles.searchModel.results[fullTestPath("css/foo.css")]).toBeUndefined();
-                        expect(FindInFiles.searchModel.results[fullTestPath("newcss/foo.css")]).toEqual(oldResults[fullTestPath("css/foo.css")]);
-                        expect(FindInFiles.searchModel.countFilesMatches()).toEqual({files: 3, matches: 14});
-                        expect(wasQuickChange).toBeFalsy();
-                    });
-                });
-            });
-
-            describe("when in-memory document changes", function () {
-                it("should update the results when a matching line is added, updating line numbers and adding the match", function () {
-                    waitms(2000);
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: fullTestPath("foo.html") }));
-                    });
-                    runs(function () {
-                        var doc = DocumentManager.getOpenDocumentForPath(fullTestPath("foo.html")),
-                            i;
-                        expect(doc).toBeTruthy();
-
-                        // Insert another line containing "foo" immediately above the second "foo" match.
-                        doc.replaceRange("this is a foo instance\n", {line: 5, ch: 0});
-
-                        // This should update synchronously.
-                        expect(gotChange).toBe(true);
-
-                        var oldFileResults = oldResults[fullTestPath("foo.html")],
-                            newFileResults = FindInFiles.searchModel.results[fullTestPath("foo.html")];
-
-                        // First match should be unchanged.
-                        expect(newFileResults.matches[0]).toEqual(oldFileResults.matches[0]);
-
-                        // Next match should be the new match. We just check the offsets here, not everything in the match record.
-                        expect(newFileResults.matches[1].start).toEqual({line: 5, ch: 10});
-                        expect(newFileResults.matches[1].end).toEqual({line: 5, ch: 13});
-
-                        // Rest of the matches should have had their lines adjusted.
-                        for (i = 2; i < newFileResults.matches.length; i++) {
-                            var newMatch = newFileResults.matches[i],
-                                oldMatch = oldFileResults.matches[i - 1];
-                            expect(newMatch.start).toEqual({line: oldMatch.start.line + 1, ch: oldMatch.start.ch});
-                            expect(newMatch.end).toEqual({line: oldMatch.end.line + 1, ch: oldMatch.end.ch});
-                        }
-
-                        // There should be one new match.
-                        expect(FindInFiles.searchModel.countFilesMatches()).toEqual({files: 3, matches: 15});
-
-                        // Make sure the model is adding the flag that will make the view debounce changes.
-                        expect(wasQuickChange).toBeTruthy();
-                    });
-                });
-
-                it("should update the results when a matching line is deleted, updating line numbers and removing the match", function () {
-                    waitms(2000);
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: fullTestPath("foo.html") }));
-                    });
-                    runs(function () {
-                        var doc = DocumentManager.getOpenDocumentForPath(fullTestPath("foo.html")),
-                            i;
-                        expect(doc).toBeTruthy();
-
-                        // Remove the second "foo" match.
-                        doc.replaceRange("", {line: 5, ch: 0}, {line: 6, ch: 0});
-
-                        // This should update synchronously.
-                        expect(gotChange).toBe(true);
-
-                        var oldFileResults = oldResults[fullTestPath("foo.html")],
-                            newFileResults = FindInFiles.searchModel.results[fullTestPath("foo.html")];
-
-                        // First match should be unchanged.
-                        expect(newFileResults.matches[0]).toEqual(oldFileResults.matches[0]);
-
-                        // Second match should be deleted. The rest of the matches should have their lines adjusted.
-                        for (i = 1; i < newFileResults.matches.length; i++) {
-                            var newMatch = newFileResults.matches[i],
-                                oldMatch = oldFileResults.matches[i + 1];
-                            expect(newMatch.start).toEqual({line: oldMatch.start.line - 1, ch: oldMatch.start.ch});
-                            expect(newMatch.end).toEqual({line: oldMatch.end.line - 1, ch: oldMatch.end.ch});
-                        }
-
-                        // There should be one fewer match.
-                        expect(FindInFiles.searchModel.countFilesMatches()).toEqual({files: 3, matches: 13});
-
-                        // Make sure the model is adding the flag that will make the view debounce changes.
-                        expect(wasQuickChange).toBeTruthy();
-                    });
-                });
-
-                it("should replace matches in a portion of the document that was edited to include a new match", function () {
-                    waitms(2000);
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: fullTestPath("foo.html") }));
-                    });
-                    runs(function () {
-                        var doc = DocumentManager.getOpenDocumentForPath(fullTestPath("foo.html")),
-                            i;
-                        expect(doc).toBeTruthy();
-
-                        // Replace the second and third foo matches (on two adjacent lines) with a single foo match on a single line.
-                        doc.replaceRange("this is a new foo match\n", {line: 5, ch: 0}, {line: 7, ch: 0});
-
-                        // This should update synchronously.
-                        expect(gotChange).toBe(true);
-
-                        var oldFileResults = oldResults[fullTestPath("foo.html")],
-                            newFileResults = FindInFiles.searchModel.results[fullTestPath("foo.html")];
-
-                        // First match should be unchanged.
-                        expect(newFileResults.matches[0]).toEqual(oldFileResults.matches[0]);
-
-                        // Second match should be changed to reflect the new position.
-                        expect(newFileResults.matches[1].start).toEqual({line: 5, ch: 14});
-                        expect(newFileResults.matches[1].end).toEqual({line: 5, ch: 17});
-
-                        // Third match should be deleted. The rest of the matches should have their lines adjusted.
-                        for (i = 2; i < newFileResults.matches.length; i++) {
-                            var newMatch = newFileResults.matches[i],
-                                oldMatch = oldFileResults.matches[i + 1];
-                            expect(newMatch.start).toEqual({line: oldMatch.start.line - 1, ch: oldMatch.start.ch});
-                            expect(newMatch.end).toEqual({line: oldMatch.end.line - 1, ch: oldMatch.end.ch});
-                        }
-
-                        // There should be one fewer match.
-                        expect(FindInFiles.searchModel.countFilesMatches()).toEqual({files: 3, matches: 13});
-
-                        // Make sure the model is adding the flag that will make the view debounce changes.
-                        expect(wasQuickChange).toBeTruthy();
-                    });
-                });
-
-                it("should completely remove the document from the results list if all matches in the document are deleted", function () {
-                    waitms(2000);
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: fullTestPath("foo.html") }));
-                    });
-                    runs(function () {
-                        var doc = DocumentManager.getOpenDocumentForPath(fullTestPath("foo.html"));
-                        expect(doc).toBeTruthy();
-
-                        // Replace all matches and check that the entire file was removed from the results list.
-                        doc.replaceRange("this will not match", {line: 4, ch: 0}, {line: 18, ch: 0});
-
-                        // This should update synchronously.
-                        expect(gotChange).toBe(true);
-                        expect(FindInFiles.searchModel.results[fullTestPath("foo.html")]).toBeUndefined();
-
-                        // There should be one fewer file and the matches for that file should be gone.
-                        expect(FindInFiles.searchModel.countFilesMatches()).toEqual({files: 2, matches: 7});
-
-                        // Make sure the model is adding the flag that will make the view debounce changes.
-                        expect(wasQuickChange).toBeTruthy();
-                    });
-                });
-            });
-
-            // Unfortunately, we can't easily mock file changes, so we just do them in a copy of the project.
-            // This set of tests isn't as thorough as it could be, because it's difficult to perform file
-            // ops that will exercise all possible scenarios of change events (e.g. change events with
-            // both added and removed files), and conversely it's difficult to mock all the filesystem stuff
-            // without doing a bunch of work. So this is really just a set of basic sanity tests to make
-            // sure that stuff being refactored between the change handler and the model doesn't break
-            // basic update functionality.
-            describe("when on-disk file or folder changes", function () {
-                it("should add matches for a new file", function () {
-                    var newFilePath;
-                    runs(function () {
-                        newFilePath = fullTestPath("newfoo.html");
-                        expect(FindInFiles.searchModel.results[newFilePath]).toBeFalsy();
-                        waitsForDone(promisify(FileSystem.getFileForPath(newFilePath), "write", "this is a new foo match\n"), "add new file");
-                    });
-                    waitsFor(function () { return gotChange; }, "model change event");
-                    runs(function () {
-                        var newFileResults = FindInFiles.searchModel.results[newFilePath];
-                        expect(newFileResults).toBeTruthy();
-                        expect(newFileResults.matches.length).toBe(1);
-                        expect(newFileResults.matches[0].start).toEqual({line: 0, ch: 14});
-                        expect(newFileResults.matches[0].end).toEqual({line: 0, ch: 17});
-
-                        // There should be one new file and match.
-                        expect(FindInFiles.searchModel.countFilesMatches()).toEqual({files: 4, matches: 15});
-                    });
-                });
-
-                it("should remove matches for a deleted file", function () {
-                    runs(function () {
-                        expect(FindInFiles.searchModel.results[fullTestPath("foo.html")]).toBeTruthy();
-                        waitsForDone(promisify(FileSystem.getFileForPath(fullTestPath("foo.html")), "unlink"), "delete file");
-                    });
-                    waitsFor(function () { return gotChange; }, "model change event");
-                    runs(function () {
-                        expect(FindInFiles.searchModel.results[fullTestPath("foo.html")]).toBeFalsy();
-
-                        // There should be one fewer file and the matches should be removed.
-                        expect(FindInFiles.searchModel.countFilesMatches()).toEqual({files: 2, matches: 7});
-                    });
-                });
-
-                it("should remove matches for a deleted folder", function () {
-                    runs(function () {
-                        expect(FindInFiles.searchModel.results[fullTestPath("css/foo.css")]).toBeTruthy();
-                        waitsForDone(promisify(FileSystem.getFileForPath(fullTestPath("css")), "unlink"), "delete folder");
-                    });
-                    waitsFor(function () { return gotChange; }, "model change event");
-                    runs(function () {
-                        expect(FindInFiles.searchModel.results[fullTestPath("css/foo.css")]).toBeFalsy();
-
-                        // There should be one fewer file and the matches should be removed.
-                        expect(FindInFiles.searchModel.countFilesMatches()).toEqual({files: 2, matches: 11});
-                    });
-                });
-            });
-        });
-
-        describe("Replace", function () {
-            function expectProjectToMatchKnownGood(kgFolder, lineEndings, filesToSkip) {
-                runs(function () {
-                    var testRootPath = ProjectManager.getProjectRoot().fullPath,
-                        kgRootPath = SpecRunnerUtils.getTestPath("/spec/FindReplace-known-goods/" + kgFolder + "/");
-
-                    function compareKnownGoodToTestFile(kgContents, kgFilePath) {
-                        var testFilePath = testRootPath + kgFilePath.slice(kgRootPath.length);
-                        if (!filesToSkip || filesToSkip.indexOf(testFilePath) === -1) {
-                            return promisify(FileSystem.getFileForPath(testFilePath), "read").then(function (testContents) {
-                                if (lineEndings) {
-                                    kgContents = FileUtils.translateLineEndings(kgContents, lineEndings);
-                                }
-                                expect(testContents).toEqual(kgContents);
-                            });
-                        }
-                    }
-
-                    waitsForDone(visitAndProcessFiles(kgRootPath, compareKnownGoodToTestFile), "project comparison done");
-                });
-            }
-
-            // Does a standard test for files on disk: search, replace, and check that files on disk match.
-            // Options:
-            //      knownGoodFolder: name of folder containing known goods to match to project files on disk
-            //      lineEndings: optional, one of the FileUtils.LINE_ENDINGS_* constants
-            //          - if specified, files on disk are expected to have these line endings
-            //      uncheckMatches: optional array of {file: string, index: number} items to uncheck; if
-            //          index unspecified, will uncheck all matches in file
-            function doBasicTest(options) {
-                doSearch(options);
-
-                runs(function () {
-                    if (options.uncheckMatches) {
-                        options.uncheckMatches.forEach(function (matchToUncheck) {
-                            var matches = searchResults[testPath + matchToUncheck.file].matches;
-                            if (matchToUncheck.index) {
-                                matches[matchToUncheck.index].isChecked = false;
-                            } else {
-                                matches.forEach(function (match) {
-                                    match.isChecked = false;
-                                });
-                            }
-                        });
-                    }
-                    waitsForDone(doReplace(options), "finish replacement");
-                });
-                expectProjectToMatchKnownGood(options.knownGoodFolder, options.lineEndings);
-            }
-
-            // Like doBasicTest, but expects some files to have specific errors.
-            // Options: same as doBasicTest, plus:
-            //      test: optional function (which must contain one or more runs blocks) to run between
-            //          search and replace
-            //      errors: array of errors expected to occur (in the same format as performReplacement() returns)
-            function doTestWithErrors(options) {
-                var done = false;
-
-                doSearch(options);
-
-                if (options.test) {
-                    // The test function *must* contain one or more runs blocks.
-                    options.test();
-                }
-
-                runs(function () {
-                    doReplace(options)
-                        .then(function () {
-                            expect("should fail due to error").toBe(true);
-                            done = true;
-                        }, function (errors) {
-                            expect(errors).toEqual(options.errors);
-                            done = true;
-                        });
-                });
-                waitsFor(function () { return done; }, 1000, "finish replacement");
-                expectProjectToMatchKnownGood(options.knownGoodFolder, options.lineEndings);
-            }
-
-            function expectInMemoryFiles(options) {
-                runs(function () {
-                    waitsForDone(Async.doInParallel(options.inMemoryFiles, function (filePath) {
-                        var fullPath;
-
-                        // If this is a full file path (as would be the case for an external file), handle it specially.
-                        if (typeof filePath === "object" && filePath.fullPath) {
-                            fullPath = filePath.fullPath;
-                            filePath = "/" + FileUtils.getBaseName(fullPath);
-                        } else {
-                            fullPath = testPath + filePath;
-                        }
-
-                        // Check that the document open in memory was changed and matches the expected replaced version of that file.
-                        var doc = DocumentManager.getOpenDocumentForPath(fullPath);
-                        expect(doc).toBeTruthy();
-                        expect(doc.isDirty).toBe(true);
-
-                        var kgPath = SpecRunnerUtils.getTestPath("/spec/FindReplace-known-goods/" + options.inMemoryKGFolder + filePath),
-                            kgFile = FileSystem.getFileForPath(kgPath);
-                        return promisify(kgFile, "read").then(function (contents) {
-                            expect(doc.getText(true)).toEqual(contents);
-                        });
-                    }), "check in memory file contents");
-                });
-            }
-
-            // Like doBasicTest, but expects one or more files to be open in memory and the replacements to happen there.
-            // Options: same as doBasicTest, plus:
-            //      inMemoryFiles: array of project-relative paths (each starting with "/") to files that should be open in memory
-            //      inMemoryKGFolder: folder containing known goods to compare each of the inMemoryFiles to
-            function doInMemoryTest(options) {
-                // Like the basic test, we expect everything on disk to match the kgFolder (which means the file open in memory
-                // should *not* have changed on disk yet).
-                doBasicTest(options);
-                expectInMemoryFiles(options);
-            }
-
-            afterEach(function () {
-                runs(function () {
-                    waitsForDone(CommandManager.execute(Commands.FILE_CLOSE_ALL, { _forceClose: true }), "close all files");
-                });
-                waitms(2000);
-            });
-
-            describe("Engine", function () {
-                it("should replace all instances of a simple string in a project on disk case-insensitively", function () {
-                    openTestProjectCopy(defaultSourcePath);
-                    doBasicTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive"
-                    });
-                });
-
-                it("should replace all instances of a simple string in a project on disk case-sensitively", function () {
-                    openTestProjectCopy(defaultSourcePath);
-                    doBasicTest({
-                        queryInfo: {query: "foo", isCaseSensitive: true},
-                        numMatches: 9,
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-sensitive"
-                    });
-                });
-
-                it("should replace all instances of a regexp in a project on disk case-insensitively with a simple replace string", function () {
-                    openTestProjectCopy(defaultSourcePath);
-                    doBasicTest({
-                        queryInfo: {query: "\\b[a-z]{3}\\b", isRegexp: true},
-                        numMatches: 33,
-                        replaceText: "CHANGED",
-                        knownGoodFolder: "regexp-case-insensitive"
-                    });
-                });
-
-                it("should replace all instances of a regexp that spans multiple lines in a project on disk", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    // This query should find each rule in the CSS file (but not in the JS file since there's more than one line
-                    // between each pair of braces).
-                    doBasicTest({
-                        queryInfo: {query: "\\{\\n[^\\n]*\\n\\}", isRegexp: true},
-                        numMatches: 4,
-                        replaceText: "CHANGED",
-                        knownGoodFolder: "regexp-replace-multiline"
-                    });
-                });
-
-                it("should replace all instances of a regexp that spans multiple lines in a project in memory", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    // This query should find each rule in the CSS file (but not in the JS file since there's more than one line
-                    // between each pair of braces).
-                    doInMemoryTest({
-                        queryInfo: {query: "\\{\\n[^\\n]*\\n\\}", isRegexp: true},
-                        numMatches: 4,
-                        replaceText: "CHANGED",
-                        knownGoodFolder: "unchanged",
-                        forceFilesOpen: true,
-                        inMemoryFiles: ["/css/foo.css"],
-                        inMemoryKGFolder: "regexp-replace-multiline"
-                    });
-                });
-
-                it("should replace all instances of a regexp that spans multiple lines in a project on disk when the last line is a partial match", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    // This query should match from the open brace through to (and including) the first colon of each rule in the
-                    // CSS file.
-                    doBasicTest({
-                        queryInfo: {query: "\\{\\n[^:]+:", isRegexp: true},
-                        numMatches: 4,
-                        replaceText: "CHANGED",
-                        knownGoodFolder: "regexp-replace-multiline-partial"
-                    });
-                });
-
-                it("should replace all instances of a regexp that spans multiple lines in a project in memory when the last line is a partial match", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    // This query should match from the open brace through to (and including) the first colon of each rule in the
-                    // CSS file.
-                    doInMemoryTest({
-                        queryInfo: {query: "\\{\\n[^:]+:", isRegexp: true},
-                        numMatches: 4,
-                        replaceText: "CHANGED",
-                        knownGoodFolder: "unchanged",
-                        forceFilesOpen: true,
-                        inMemoryFiles: ["/css/foo.css"],
-                        inMemoryKGFolder: "regexp-replace-multiline-partial"
-                    });
-                });
-
-                it("should replace all instances of a regexp in a project on disk case-sensitively with a simple replace string", function () {
-                    openTestProjectCopy(defaultSourcePath);
-                    doBasicTest({
-                        queryInfo: {query: "\\b[a-z]{3}\\b", isRegexp: true, isCaseSensitive: true},
-                        numMatches: 25,
-                        replaceText: "CHANGED",
-                        knownGoodFolder: "regexp-case-sensitive"
-                    });
-                });
-
-                it("should replace instances of a regexp with a $-substitution on disk", function () {
-                    openTestProjectCopy(defaultSourcePath);
-                    doBasicTest({
-                        queryInfo: {query: "\\b([a-z]{3})\\b", isRegexp: true},
-                        numMatches: 33,
-                        replaceText: "[$1]",
-                        knownGoodFolder: "regexp-dollar-replace"
-                    });
-                });
-
-                it("should replace instances of a regexp with a $-substitution in in-memory files", function () {
-                    // This test case is necessary because the in-memory case goes through a separate code path before it deals with
-                    // the replace text.
-                    openTestProjectCopy(defaultSourcePath);
-
-                    doInMemoryTest({
-                        queryInfo: {query: "\\b([a-z]{3})\\b", isRegexp: true},
-                        numMatches: 33,
-                        replaceText: "[$1]",
-                        knownGoodFolder: "unchanged",
-                        forceFilesOpen: true,
-                        inMemoryFiles: ["/css/foo.css", "/foo.html", "/foo.js"],
-                        inMemoryKGFolder: "regexp-dollar-replace"
-                    });
-                });
-
-                it("should replace instances of regexp with 0-length matches on disk", function () {
-                    openTestProjectCopy(defaultSourcePath);
-                    doBasicTest({
-                        queryInfo: {query: "^", isRegexp: true},
-                        numMatches: 55,
-                        replaceText: "CHANGED",
-                        knownGoodFolder: "regexp-zero-length"
-                    });
-                });
-
-                it("should replace instances of regexp with 0-length matches in memory", function () {
-                    openTestProjectCopy(defaultSourcePath);
-                    doInMemoryTest({
-                        queryInfo: {query: "^", isRegexp: true},
-                        numMatches: 55,
-                        replaceText: "CHANGED",
-                        knownGoodFolder: "unchanged",
-                        forceFilesOpen: true,
-                        inMemoryFiles: ["/css/foo.css", "/foo.html", "/foo.js"],
-                        inMemoryKGFolder: "regexp-zero-length"
-                    });
-                });
-
-                it("should replace instances of a string in a project respecting CRLF line endings", function () {
-                    openTestProjectCopy(defaultSourcePath, FileUtils.LINE_ENDINGS_CRLF);
-                    doBasicTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive",
-                        lineEndings: FileUtils.LINE_ENDINGS_CRLF
-                    });
-                });
-
-                it("should replace instances of a string in a project respecting LF line endings", function () {
-                    openTestProjectCopy(defaultSourcePath, FileUtils.LINE_ENDINGS_LF);
-                    doBasicTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive",
-                        lineEndings: FileUtils.LINE_ENDINGS_LF
-                    });
-                });
-
-                it("should not replace unchecked matches on disk", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    doBasicTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        uncheckMatches: [{file: "/css/foo.css"}],
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive-except-foo.css"
-                    });
-                });
-
-                it("should do all in-memory replacements synchronously, so user can't accidentally edit document after start of replace process", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    // Open two of the documents we want to replace in memory.
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: testPath + "/css/foo.css" }), "opening document");
-                    });
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: testPath + "/foo.js" }), "opening document");
-                    });
-
-                    // We can't use expectInMemoryFiles(), since this test requires everything to happen fully synchronously
-                    // (no file reads) once the replace has started. So we read the files here.
-                    var kgFileContents = {};
-                    runs(function () {
-                        var kgPath = SpecRunnerUtils.getTestPath("/spec/FindReplace-known-goods/simple-case-insensitive");
-                        waitsForDone(visitAndProcessFiles(kgPath, function (contents, fullPath) {
-                            // Translate line endings to in-memory document style (always LF)
-                            kgFileContents[fullPath.slice(kgPath.length)] = FileUtils.translateLineEndings(contents, FileUtils.LINE_ENDINGS_LF);
-                        }), "reading known good");
-                    });
-
-                    doSearch({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar"
-                    });
-
-                    runs(function () {
-                        // Start the replace, but don't wait for it to complete. Since the in-memory replacements should occur
-                        // synchronously, the in-memory documents should have already been changed. This means we don't have to
-                        // worry about detecting changes in documents once the replace starts. (If the user had  changed
-                        // the document after the search but before the replace started, we would have already closed the panel,
-                        // preventing the user from doing a replace.)
-                        var promise = FindInFiles.doReplace(searchResults, "bar");
-
-                        // Check the in-memory contents against the known goods.
-                        ["/css/foo.css", "/foo.js"].forEach(function (filename) {
-                            var fullPath = testPath + filename,
-                                doc = DocumentManager.getOpenDocumentForPath(fullPath);
-                            expect(doc).toBeTruthy();
-                            expect(doc.isDirty).toBe(true);
-                            expect(doc.getText()).toEqual(kgFileContents[filename]);
-                        });
-
-                        // Finish the replace operation, which should go ahead and do the file on disk.
-                        waitsForDone(promise);
-                    });
-
-                    runs(function () {
-                        // Now the file on disk should have been replaced too.
-                        waitsForDone(promisify(FileSystem.getFileForPath(testPath + "/foo.html"), "read").then(function (contents) {
-                            expect(FileUtils.translateLineEndings(contents, FileUtils.LINE_ENDINGS_LF)).toEqual(kgFileContents["/foo.html"]);
-                        }), "checking known good");
-                    });
-                });
-
-                it("should return an error and not do the replacement in files that have changed on disk since the search", function () {
-                    openTestProjectCopy(defaultSourcePath);
-                    doTestWithErrors({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "changed-file",
-                        test: function () {
-                            // Wait for one second to make sure that the changed file gets an updated timestamp.
-                            // TODO: this seems like a FileSystem issue - we don't get timestamp changes with a resolution
-                            // of less than one second.
-                            waits(1000);
-
-                            runs(function () {
-                                // Clone the results so we don't use the version that's auto-updated by FindInFiles when we modify the file
-                                // on disk. This case might not usually come up in the real UI if we always guarantee that the results list will
-                                // be auto-updated, but we want to make sure there's no edge case where we missed an update and still clobber the
-                                // file on disk anyway.
-                                searchResults = _.cloneDeep(searchResults);
-                                waitsForDone(promisify(FileSystem.getFileForPath(testPath + "/css/foo.css"), "write", "/* changed content */"), "modify file");
-                            });
-                        },
-                        errors: [{item: testPath + "/css/foo.css", error: FindUtils.ERROR_FILE_CHANGED}]
-                    });
-                });
-
-                it("should return an error if a write fails", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    // Return a fake error when we try to write to the CSS file. (Note that this is spying on the test window's File module.)
-                    var writeSpy = spyOn(File.prototype, "write").andCallFake(function (data, options, callback) {
-                        if (typeof options === "function") {
-                            callback = options;
-                        } else {
-                            callback = callback || function () {};
-                        }
-                        if (this.fullPath === testPath + "/css/foo.css") {
-                            callback(FileSystemError.NOT_WRITABLE);
-                        } else {
-                            return writeSpy.originalValue.apply(this, arguments);
-                        }
-                    });
-
-                    doTestWithErrors({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive-except-foo.css",
-                        errors: [{item: testPath + "/css/foo.css", error: FileSystemError.NOT_WRITABLE}]
-                    });
-                });
-
-                it("should return an error if a match timestamp doesn't match an in-memory document timestamp", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: testPath + "/css/foo.css" }), "opening document");
-                    });
-
-                    doTestWithErrors({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive-except-foo.css",
-                        test: function () {
-                            runs(function () {
-                                // Clone the results so we don't use the version that's auto-updated by FindInFiles when we modify the file
-                                // on disk. This case might not usually come up in the real UI if we always guarantee that the results list will
-                                // be auto-updated, but we want to make sure there's no edge case where we missed an update and still clobber the
-                                // file on disk anyway.
-                                searchResults = _.cloneDeep(searchResults);
-                                var oldTimestamp = searchResults[testPath + "/css/foo.css"].timestamp;
-                                searchResults[testPath + "/css/foo.css"].timestamp = new Date(oldTimestamp.getTime() - 5000);
-                            });
-                        },
-                        errors: [{item: testPath + "/css/foo.css", error: FindUtils.ERROR_FILE_CHANGED}]
-                    });
-                });
-
-                it("should do the replacement in memory for a file open in an Editor in the working set", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, {fullPath: testPath + "/css/foo.css"}), "add file to working set");
-                    });
-
-                    doInMemoryTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive-except-foo.css",
-                        inMemoryFiles: ["/css/foo.css"],
-                        inMemoryKGFolder: "simple-case-insensitive"
-                    });
-                });
-
-                it("should do the search/replace in the current document content for a dirty in-memory document", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    var options = {
-                        queryInfo: {query: "foo"},
-                        numMatches: 15,
-                        replaceText: "bar",
-                        inMemoryFiles: ["/css/foo.css"],
-                        inMemoryKGFolder: "simple-case-insensitive-modified"
-                    };
-
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, {fullPath: testPath + "/css/foo.css"}), "add file to working set");
-                    });
-                    runs(function () {
-                        var doc = DocumentManager.getOpenDocumentForPath(testPath + "/css/foo.css");
-                        expect(doc).toBeTruthy();
-                        doc.replaceRange("/* added a foo line */\n", {line: 0, ch: 0});
-                    });
-                    doSearch(options);
-                    runs(function () {
-                        waitsForDone(doReplace(options), "replace done");
-                    });
-                    expectInMemoryFiles(options);
-                    expectProjectToMatchKnownGood("simple-case-insensitive-modified", null, [testPath + "/css/foo.css"]);
-                });
-
-                it("should do the replacement in memory for a file open in an Editor that's not in the working set", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.FILE_OPEN, {fullPath: testPath + "/css/foo.css"}), "open file");
-                    });
-
-                    doInMemoryTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive-except-foo.css",
-                        inMemoryFiles: ["/css/foo.css"],
-                        inMemoryKGFolder: "simple-case-insensitive"
-                    });
-                });
-
-                it("should do the replacement in memory for a file that's in the working set but not yet open in an editor", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    runs(function () {
-                        MainViewManager.addToWorkingSet(MainViewManager.ACTIVE_PANE, FileSystem.getFileForPath(testPath + "/css/foo.css"));
-                    });
-
-                    doInMemoryTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive-except-foo.css",
-                        inMemoryFiles: ["/css/foo.css"],
-                        inMemoryKGFolder: "simple-case-insensitive"
-                    });
-                });
-
-                it("should open the document in an editor and do the replacement there if the document is open but not in an Editor", function () {
-                    var doc, openFilePath;
-                    openTestProjectCopy(defaultSourcePath);
-
-                    runs(function () {
-                        openFilePath = testPath + "/css/foo.css";
-                        waitsForDone(DocumentManager.getDocumentForPath(openFilePath).done(function (d) {
-                            doc = d;
-                            doc.addRef();
-                        }), "get document");
-                    });
-
-                    doInMemoryTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive-except-foo.css",
-                        inMemoryFiles: ["/css/foo.css"],
-                        inMemoryKGFolder: "simple-case-insensitive"
-                    });
-
-                    runs(function () {
-                        var workingSet = MainViewManager.getWorkingSet(MainViewManager.ALL_PANES);
-                        expect(workingSet.some(function (file) { return file.fullPath === openFilePath; })).toBe(true);
-                        doc.releaseRef();
-                    });
-                });
-
-                it("should open files and do all replacements in memory if forceFilesOpen is true", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    doInMemoryTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "unchanged",
-                        forceFilesOpen: true,
-                        inMemoryFiles: ["/css/foo.css", "/foo.html", "/foo.js"],
-                        inMemoryKGFolder: "simple-case-insensitive"
-                    });
-                });
-
-                it("should not perform unchecked matches in memory", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    doInMemoryTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        uncheckMatches: [{file: "/css/foo.css", index: 1}, {file: "/foo.html", index: 3}],
-                        replaceText: "bar",
-                        knownGoodFolder: "unchanged",
-                        forceFilesOpen: true,
-                        inMemoryFiles: ["/css/foo.css", "/foo.html", "/foo.js"],
-                        inMemoryKGFolder: "simple-case-insensitive-unchecked"
-                    });
-                });
-
-                it("should not perform unchecked matches on disk", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    doBasicTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        uncheckMatches: [{file: "/css/foo.css", index: 1}, {file: "/foo.html", index: 3}],
-                        replaceText: "bar",
-                        knownGoodFolder: "simple-case-insensitive-unchecked"
-                    });
-                });
-
-                it("should select the first modified file in the working set if replacements are done in memory and current editor wasn't affected", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, {fullPath: testPath + "/bar.txt"}), "open file");
-                    });
-
-                    doInMemoryTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "unchanged",
-                        forceFilesOpen: true,
-                        inMemoryFiles: ["/css/foo.css", "/foo.html", "/foo.js"],
-                        inMemoryKGFolder: "simple-case-insensitive"
-                    });
-
-                    runs(function () {
-                        var expectedFile = testPath + "/foo.html";
-                        expect(DocumentManager.getCurrentDocument().file.fullPath).toBe(expectedFile);
-                        expect(MainViewManager.findInWorkingSet(MainViewManager.ACTIVE_PANE, expectedFile)).not.toBe(-1);
-                    });
-                });
-
-                it("should select the first modified file in the working set if replacements are done in memory and no editor was open", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    var testFiles = ["/css/foo.css", "/foo.html", "/foo.js"];
-
-                    doInMemoryTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        replaceText: "bar",
-                        knownGoodFolder: "unchanged",
-                        forceFilesOpen: true,
-                        inMemoryFiles: testFiles,
-                        inMemoryKGFolder: "simple-case-insensitive"
-                    });
-
-
-                    runs(function () {
-                        // since nothing was opened prior to doing the
-                        //  replacements then the first file modified will be opened.
-                        // This may not be the first item in the array above
-                        //  since the files are sorted differently in performReplacements
-                        //  and the replace is performed asynchronously.
-                        // So, just ensure that *something* was opened
-                        expect(DocumentManager.getCurrentDocument().file.fullPath).toBeTruthy();
-
-                        testFiles.forEach(function (relPath) {
-                            expect(MainViewManager.findInWorkingSet(MainViewManager.ACTIVE_PANE, testPath + relPath)).not.toBe(-1);
-                        });
-                    });
-                });
-
-                it("should select the first modified file in the working set if replacements are done in memory and there were no matches checked for current editor", function () {
-                    openTestProjectCopy(defaultSourcePath);
-
-                    runs(function () {
-                        waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, {fullPath: testPath + "/css/foo.css"}), "open file");
-                    });
-
-                    doInMemoryTest({
-                        queryInfo: {query: "foo"},
-                        numMatches: 14,
-                        uncheckMatches: [{file: "/css/foo.css"}],
-                        replaceText: "bar",
-                        knownGoodFolder: "unchanged",
-                        forceFilesOpen: true,
-                        inMemoryFiles: ["/foo.html", "/foo.js"],
-                        inMemoryKGFolder: "simple-case-insensitive-except-foo.css"
-                    });
-
-                    runs(function () {
-                        expect(DocumentManager.getCurrentDocument().file.fullPath).toEqual(testPath + "/foo.html");
-                    });
-                });
-            });
-
-            describe("UI", function () {
-                function executeReplace(findText, replaceText, fromKeyboard) {
-                    runs(function () {
-                        FindInFiles._searchDone = false;
-                        FindInFiles._replaceDone = false;
-                        $("#find-what").val(findText).trigger("input");
-                        $("#replace-with").val(replaceText).trigger("input");
-                        if (fromKeyboard) {
-                            SpecRunnerUtils.simulateKeyEvent(KeyEvent.DOM_VK_RETURN, "keydown", $("#replace-with").get(0));
-                        } else {
-                            $("#replace-batch").click();
-                        }
-                    });
-                }
-
-                function showSearchResults(findText, replaceText, fromKeyboard) {
-                    openTestProjectCopy(defaultSourcePath);
-                    openSearchBar(null, true);
-                    executeReplace(findText, replaceText, fromKeyboard);
-                    waitsFor(function () {
-                        return FindInFiles._searchDone;
-                    }, "search finished");
-                }
-
-                afterEach(function () {
-                    closeSearchBar();
-                });
-
-                describe("Replace in Files Bar", function () {
-                    it("should only show a Replace All button", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        openSearchBar(null, true);
-                        runs(function () {
-                            expect($("#replace-yes").length).toBe(0);
-                            expect($("#replace-batch").length).toBe(1);
-                        });
-                    });
-
-                    it("should disable the Replace button if query is empty", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        openSearchBar(null, true);
-                        runs(function () {
-                            $("#find-what").val("").trigger("input");
-                            expect($("#replace-batch").is(":disabled")).toBe(true);
-                        });
-                    });
-
-                    it("should enable the Replace button if the query is a non-empty string", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        openSearchBar(null, true);
-                        runs(function () {
-                            $("#find-what").val("my query").trigger("input");
-                            expect($("#replace-batch").is(":disabled")).toBe(false);
-                        });
-                    });
-
-                    it("should disable the Replace button if query is an invalid regexp", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        openSearchBar(null, true);
-                        runs(function () {
-                            $("#find-regexp").click();
-                            $("#find-what").val("[invalid").trigger("input");
-                            expect($("#replace-batch").is(":disabled")).toBe(true);
-                        });
-                    });
-
-                    it("should enable the Replace button if query is a valid regexp", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        openSearchBar(null, true);
-                        runs(function () {
-                            $("#find-regexp").click();
-                            $("#find-what").val("[valid]").trigger("input");
-                            expect($("#replace-batch").is(":disabled")).toBe(false);
-                        });
-                    });
-
-                    it("should start with focus in Find, and set focus to the Replace field when the user hits enter in the Find field", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        openSearchBar(null, true);
-                        runs(function () {
-                            // For some reason using $().is(":focus") here is flaky.
-                            expect(testWindow.document.activeElement).toBe($("#find-what").get(0));
-                            SpecRunnerUtils.simulateKeyEvent(KeyEvent.DOM_VK_RETURN, "keydown", $("#find-what").get(0));
-                            expect(testWindow.document.activeElement).toBe($("#replace-with").get(0));
-                        });
-                    });
-                });
-
-                describe("Full workflow", function () {
-                    it("should prepopulate the find bar with selected text", function () {
-                        var doc, editor;
-
-                        openTestProjectCopy(defaultSourcePath);
-                        runs(function () {
-                            waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: testPath + "/foo.html" }), "open file");
-                        });
-                        runs(function () {
-                            doc = DocumentManager.getOpenDocumentForPath(testPath + "/foo.html");
-                            expect(doc).toBeTruthy();
-                            MainViewManager._edit(MainViewManager.ACTIVE_PANE, doc);
-                            editor = doc._masterEditor;
-                            expect(editor).toBeTruthy();
-                            editor.setSelection({line: 4, ch: 7}, {line: 4, ch: 10});
-                        });
-
-                        openSearchBar(null);
-                        runs(function () {
-                            expect($("#find-what").val()).toBe("Foo");
-                        });
-                        waitsForDone(CommandManager.execute(Commands.FILE_CLOSE_ALL), "closing all files");
-                    });
-
-                    it("should prepopulate the find bar with only first line of selected text", function () {
-                        var doc, editor;
-
-                        openTestProjectCopy(defaultSourcePath);
-                        runs(function () {
-                            waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: testPath + "/foo.html" }), "open file");
-                        });
-                        runs(function () {
-                            doc = DocumentManager.getOpenDocumentForPath(testPath + "/foo.html");
-                            expect(doc).toBeTruthy();
-                            MainViewManager._edit(MainViewManager.ACTIVE_PANE, doc);
-                            editor = doc._masterEditor;
-                            expect(editor).toBeTruthy();
-                            editor.setSelection({line: 4, ch: 7}, {line: 6, ch: 10});
-                        });
-
-                        openSearchBar(null);
-                        runs(function () {
-                            expect($("#find-what").val()).toBe("Foo</title>");
-                        });
-                        waitsForDone(CommandManager.execute(Commands.FILE_CLOSE_ALL), "closing all files");
-                    });
-
-                    it("should show results from the search with all checkboxes checked", function () {
-                        showSearchResults("foo", "bar");
-                        runs(function () {
-                            expect($("#find-in-files-results").length).toBe(1);
-                            expect($("#find-in-files-results .check-one").length).toBe(14);
-                            expect($("#find-in-files-results .check-one:checked").length).toBe(14);
-                        });
-                    });
-
-                    it("should do a simple search/replace all from find bar, opening results in memory, when user clicks on Replace... button", function () {
-                        showSearchResults("foo", "bar");
-                        // Click the "Replace" button in the search panel - this should kick off the replace
-                        runs(function () {
-                            $(".replace-checked").click();
-                        });
-
-                        waitsFor(function () {
-                            return FindInFiles._replaceDone;
-                        }, "replace finished");
-                        expectInMemoryFiles({
-                            inMemoryFiles: ["/css/foo.css", "/foo.html", "/foo.js"],
-                            inMemoryKGFolder: "simple-case-insensitive"
-                        });
-                    });
-
-                    it("should do a simple search/replace all from find bar, opening results in memory, when user hits Enter in Replace field", function () {
-                        showSearchResults("foo", "bar");
-                        // Click the "Replace" button in the search panel - this should kick off the replace
-                        runs(function () {
-                            $(".replace-checked").click();
-                        });
-
-                        waitsFor(function () {
-                            return FindInFiles._replaceDone;
-                        }, "replace finished");
-                        expectInMemoryFiles({
-                            inMemoryFiles: ["/css/foo.css", "/foo.html", "/foo.js"],
-                            inMemoryKGFolder: "simple-case-insensitive"
-                        });
-                    });
-
-                    it("should do a search in folder, replace all from find bar", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        var dirEntry = FileSystem.getDirectoryForPath(testPath + "/css/");
-                        openSearchBar(dirEntry, true);
-                        executeReplace("foo", "bar", true);
-
-                        waitsFor(function () {
-                            return FindInFiles._searchDone;
-                        }, "search finished");
-
-                        // Click the "Replace" button in the search panel - this should kick off the replace
-                        runs(function () {
-                            $(".replace-checked").click();
-                        });
-
-                        waitsFor(function () {
-                            return FindInFiles._replaceDone;
-                        }, "replace finished");
-                        expectInMemoryFiles({
-                            inMemoryFiles: ["/css/foo.css"],
-                            inMemoryKGFolder: "simple-case-insensitive-only-foo.css"
-                        });
-                    });
-
-                    it("should do a search in file, replace all from find bar", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        var fileEntry = FileSystem.getFileForPath(testPath + "/css/foo.css");
-                        openSearchBar(fileEntry, true);
-                        executeReplace("foo", "bar", true);
-
-                        waitsFor(function () {
-                            return FindInFiles._searchDone;
-                        }, "search finished");
-
-                        // Click the "Replace" button in the search panel - this should kick off the replace
-                        runs(function () {
-                            $(".replace-checked").click();
-                        });
-
-                        waitsFor(function () {
-                            return FindInFiles._replaceDone;
-                        }, "replace finished");
-                        expectInMemoryFiles({
-                            inMemoryFiles: ["/css/foo.css"],
-                            inMemoryKGFolder: "simple-case-insensitive-only-foo.css"
-                        });
-                    });
-
-                    it("should do a regexp search/replace from find bar", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        openSearchBar(null, true);
-                        runs(function () {
-                            $("#find-regexp").click();
-                        });
-                        executeReplace("\\b([a-z]{3})\\b", "[$1]", true);
-
-                        waitsFor(function () {
-                            return FindInFiles._searchDone;
-                        }, "search finished");
-
-                        // Click the "Replace" button in the search panel - this should kick off the replace
-                        runs(function () {
-                            $(".replace-checked").click();
-                        });
-
-                        waitsFor(function () {
-                            return FindInFiles._replaceDone;
-                        }, "replace finished");
-                        expectInMemoryFiles({
-                            inMemoryFiles: ["/css/foo.css", "/foo.html", "/foo.js"],
-                            inMemoryKGFolder: "regexp-dollar-replace"
-                        });
-                    });
-
-                    it("should do a case-sensitive search/replace from find bar", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        openSearchBar(null, true);
-                        runs(function () {
-                            $("#find-case-sensitive").click();
-                        });
-                        executeReplace("foo", "bar", true);
-
-                        waitsFor(function () {
-                            return FindInFiles._searchDone;
-                        }, "search finished");
-
-                        // Click the "Replace" button in the search panel - this should kick off the replace
-                        runs(function () {
-                            $(".replace-checked").click();
-                        });
-
-                        waitsFor(function () {
-                            return FindInFiles._replaceDone;
-                        }, "replace finished");
-                        expectInMemoryFiles({
-                            inMemoryFiles: ["/css/foo.css", "/foo.html", "/foo.js"],
-                            inMemoryKGFolder: "simple-case-sensitive"
-                        });
-                    });
-
-                    it("should warn and do changes on disk if there are changes in >20 files", function () {
-                        openTestProjectCopy(SpecRunnerUtils.getTestPath("/spec/FindReplace-test-files-large"));
-                        openSearchBar(null, true);
-                        executeReplace("foo", "bar");
-
-                        waitsFor(function () {
-                            return FindInFiles._searchDone;
-                        }, "search finished");
-
-                        // Click the "Replace" button in the search panel - this should cause the dialog to appear
-                        runs(function () {
-                            $(".replace-checked").click();
-                        });
-
-                        runs(function () {
-                            expect(FindInFiles._replaceDone).toBeFalsy();
-                        });
-
-                        var $okButton;
-                        waitsFor(function () {
-                            $okButton = $(".dialog-button[data-button-id='ok']");
-                            return !!$okButton.length;
-                        }, "dialog appearing");
-                        runs(function () {
-                            expect($okButton.length).toBe(1);
-                            expect($okButton.text()).toBe(Strings.BUTTON_REPLACE_WITHOUT_UNDO);
-                            $okButton.click();
-                        });
-
-                        waitsFor(function () {
-                            return FindInFiles._replaceDone;
-                        }, "replace finished");
-                        expectProjectToMatchKnownGood("simple-case-insensitive-large");
-                    });
-
-                    it("should not do changes on disk if Cancel is clicked in 'too many files' dialog", function () {
-                        spyOn(FindInFiles, "doReplace").andCallThrough();
-                        openTestProjectCopy(SpecRunnerUtils.getTestPath("/spec/FindReplace-test-files-large"));
-                        openSearchBar(null, true);
-                        executeReplace("foo", "bar");
-
-                        waitsFor(function () {
-                            return FindInFiles._searchDone;
-                        }, "search finished");
-
-                        // Click the "Replace" button in the search panel - this should cause the dialog to appear
-                        runs(function () {
-                            $(".replace-checked").click();
-                        });
-
-                        runs(function () {
-                            expect(FindInFiles._replaceDone).toBeFalsy();
-                        });
-
-                        var $cancelButton;
-                        waitsFor(function () {
-                            $cancelButton = $(".dialog-button[data-button-id='cancel']");
-                            return !!$cancelButton.length;
-                        });
-                        runs(function () {
-                            expect($cancelButton.length).toBe(1);
-                            $cancelButton.click();
-                        });
-
-                        waitsFor(function () {
-                            return $(".dialog-button[data-button-id='cancel']").length === 0;
-                        }, "dialog dismissed");
-                        runs(function () {
-                            expect(FindInFiles.doReplace).not.toHaveBeenCalled();
-                            // Panel should be left open.
-                            expect($("#find-in-files-results").is(":visible")).toBeTruthy();
-                        });
-                    });
-
-                    it("should do single-file Replace All in an open file in the project", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        runs(function () {
-                            waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: testPath + "/foo.js" }), "open file");
-                        });
-                        runs(function () {
-                            waitsForDone(CommandManager.execute(Commands.CMD_REPLACE), "open single-file replace bar");
-                        });
-                        waitsFor(function () {
-                            return $(".modal-bar").length === 1;
-                        }, "search bar open");
-
-                        executeReplace("foo", "bar");
-                        waitsFor(function () {
-                            return FindInFiles._searchDone;
-                        }, "search finished");
-
-                        // Click the "Replace" button in the search panel - this should kick off the replace
-                        runs(function () {
-                            $(".replace-checked").click();
-                        });
-
-                        waitsFor(function () {
-                            return FindInFiles._replaceDone;
-                        }, "replace finished");
-
-                        expectInMemoryFiles({
-                            inMemoryFiles: ["/foo.js"],
-                            inMemoryKGFolder: "simple-case-insensitive"
-                        });
-                    });
-
-                    it("should do single-file Replace All in a non-project file", function () {
-                        // Open an empty project.
-                        var blankProject = SpecRunnerUtils.getTempDirectory() + "/blank-project",
-                            externalFilePath = defaultSourcePath + "/foo.js";
-                        runs(function () {
-                            var dirEntry = FileSystem.getDirectoryForPath(blankProject);
-                            waitsForDone(promisify(dirEntry, "create"));
-                        });
-                        SpecRunnerUtils.loadProjectInTestWindow(blankProject);
-                        runs(function () {
-                            waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: externalFilePath }), "open external file");
-                        });
-                        runs(function () {
-                            waitsForDone(CommandManager.execute(Commands.CMD_REPLACE), "open single-file replace bar");
-                        });
-                        waitsFor(function () {
-                            return $(".modal-bar").length === 1;
-                        }, "search bar open");
-
-                        executeReplace("foo", "bar");
-                        waitsFor(function () {
-                            return FindInFiles._searchDone;
-                        }, "search finished");
-
-                        // Click the "Replace" button in the search panel - this should kick off the replace
-                        runs(function () {
-                            $(".replace-checked").click();
-                        });
-
-                        waitsFor(function () {
-                            return FindInFiles._replaceDone;
-                        }, "replace finished");
-
-                        expectInMemoryFiles({
-                            inMemoryFiles: [{fullPath: externalFilePath}], // pass a full file path since this is an external file
-                            inMemoryKGFolder: "simple-case-insensitive"
-                        });
-                    });
-
-                    it("should show an error dialog if errors occurred during the replacement", function () {
-                        showSearchResults("foo", "bar");
-                        runs(function () {
-                            spyOn(FindInFiles, "doReplace").andCallFake(function () {
-                                return new $.Deferred().reject([
-                                    {item: testPath + "/css/foo.css", error: FindUtils.ERROR_FILE_CHANGED},
-                                    {item: testPath + "/foo.html", error: FileSystemError.NOT_WRITABLE}
-                                ]);
-                            });
-                        });
-                        runs(function () {
-                            // This will call our mock doReplace
-                            $(".replace-checked").click();
-                        });
-
-                        var $dlg;
-                        waitsFor(function () {
-                            $dlg = $(".error-dialog");
-                            return !!$dlg.length;
-                        }, "dialog appearing");
-                        runs(function () {
-                            expect($dlg.length).toBe(1);
-
-                            // Both files should be mentioned in the dialog.
-                            var text = $dlg.find(".dialog-message").text();
-                            // Have to check this in a funny way because breakableUrl() adds a special character after the slash.
-                            expect(text.match(/css\/.*foo.css/)).not.toBe(-1);
-                            expect(text.indexOf(StringUtils.breakableUrl("foo.html"))).not.toBe(-1);
-                            $dlg.find(".dialog-button[data-button-id='ok']").click();
-                            expect($(".error-dialog").length).toBe(0);
-                        });
-                    });
-                });
-
-                // TODO: these could be split out into unit tests, but would need to be able to instantiate
-                // a SearchResultsView in the test runner window.
-                describe("Checkbox interactions", function () {
-                    it("should uncheck all checkboxes and update model when Check All is clicked while checked", function () {
-                        showSearchResults("foo", "bar");
-                        runs(function () {
-                            expect($(".check-all").is(":checked")).toBeTruthy();
-                            $(".check-all").click();
-                            expect($(".check-all").is(":checked")).toBeFalsy();
-                            expect($(".check-one:checked").length).toBe(0);
-                            expect(_.find(FindInFiles.searchModel.results, function (result) {
-                                return _.find(result.matches, function (match) { return match.isChecked; });
-                            })).toBeFalsy();
-                        });
-                    });
-
-                    it("should uncheck one checkbox and update model, unchecking the Check All checkbox", function () {
-                        showSearchResults("foo", "bar");
-                        runs(function () {
-                            $(".check-one").eq(1).click();
-                            expect($(".check-one").eq(1).is(":checked")).toBeFalsy();
-                            expect($(".check-all").is(":checked")).toBeFalsy();
-                            // In the sorting, this item should be the second match in the first file, which is foo.html
-                            var uncheckedMatch = FindInFiles.searchModel.results[testPath + "/foo.html"].matches[1];
-                            expect(uncheckedMatch.isChecked).toBe(false);
-                            // Check that all items in the model besides the unchecked one to be checked.
-                            expect(_.every(FindInFiles.searchModel.results, function (result) {
-                                return _.every(result.matches, function (match) {
-                                    if (match === uncheckedMatch) {
-                                        // This one is already expected to be unchecked.
-                                        return true;
-                                    }
-                                    return match.isChecked;
-                                });
-                            })).toBeTruthy();
-                        });
-                    });
-
-                    it("should re-check unchecked checkbox and update model after clicking Check All again", function () {
-                        showSearchResults("foo", "bar");
-                        runs(function () {
-                            $(".check-one").eq(1).click();
-                            expect($(".check-one").eq(1).is(":checked")).toBeFalsy();
-                            expect($(".check-all").is(":checked")).toBeFalsy();
-                            $(".check-all").click();
-                            expect($(".check-all").is(":checked")).toBeTruthy();
-                            expect($(".check-one:checked").length).toEqual($(".check-one").length);
-                            expect(_.every(FindInFiles.searchModel.results, function (result) {
-                                return _.every(result.matches, function (match) { return match.isChecked; });
-                            })).toBeTruthy();
-                        });
-                    });
-
-                    // TODO: checkboxes with paging
-                });
-                // Untitled documents are covered in the "Search -> Replace All in untitled document" cases above.
-
-                describe("Panel closure on changes", function () {
-                    it("should close the panel and detach listeners if a file is modified on disk", function () {
-                        showSearchResults("foo", "bar");
-                        runs(function () {
-                            expect($("#find-in-files-results").is(":visible")).toBe(true);
-                            waitsForDone(promisify(FileSystem.getFileForPath(testPath + "/foo.html"), "write", "changed content"));
-                        });
-                        runs(function () {
-                            expect($("#find-in-files-results").is(":visible")).toBe(false);
-                        });
-                    });
-
-                    it("should close the panel if a file is modified in memory", function () {
-                        openTestProjectCopy(defaultSourcePath);
-                        runs(function () {
-                            waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: testPath + "/foo.html" }), "open file");
-                        });
-                        openSearchBar(null, true);
-                        executeReplace("foo", "bar");
-                        waitsFor(function () {
-                            return FindInFiles._searchDone;
-                        }, "search finished");
-                        runs(function () {
-                            expect($("#find-in-files-results").is(":visible")).toBe(true);
-
-                            var doc = DocumentManager.getOpenDocumentForPath(testPath + "/foo.html");
-                            expect(doc).toBeTruthy();
-                            doc.replaceRange("", {line: 0, ch: 0}, {line: 1, ch: 0});
-
-                            expect($("#find-in-files-results").is(":visible")).toBe(false);
-                        });
-                    });
-
-                    it("should close the panel if a document was open and modified before the search, but then the file was closed and changes dropped", function () {
-                        var doc;
-
-                        openTestProjectCopy(defaultSourcePath);
-                        runs(function () {
-                            waitsForDone(CommandManager.execute(Commands.CMD_ADD_TO_WORKINGSET_AND_OPEN, { fullPath: testPath + "/foo.html" }), "open file");
-                        });
-                        runs(function () {
-                            doc = DocumentManager.getOpenDocumentForPath(testPath + "/foo.html");
-                            expect(doc).toBeTruthy();
-                            doc.replaceRange("", {line: 0, ch: 0}, {line: 1, ch: 0});
-                        });
-                        openSearchBar(null, true);
-                        executeReplace("foo", "bar");
-                        waitsFor(function () {
-                            return FindInFiles._searchDone;
-                        }, "search finished");
-                        runs(function () {
-                            expect($("#find-in-files-results").is(":visible")).toBe(true);
-
-                            // We have to go through the dialog workflow for closing the file without saving changes,
-                            // because the "revert" behavior only happens in that workflow (it doesn't happen if you
-                            // do forceClose, since that's only intended as a shortcut for the end of a unit test).
-                            var closePromise = CommandManager.execute(Commands.FILE_CLOSE, { file: doc.file }),
-                                $dontSaveButton = $(".dialog-button[data-button-id='dontsave']");
-                            expect($dontSaveButton.length).toBe(1);
-                            $dontSaveButton.click();
-                            waitsForDone(closePromise);
-                        });
-                        runs(function () {
-                            expect($("#find-in-files-results").is(":visible")).toBe(false);
-                        });
-                    });
-                });
-
-                describe("Disclosure Arrows", function () {
-
-                    it("should expand/collapse items when clicked", function () {
-                        showSearchResults("foo", "bar");
-                        runs(function () {
-                            $(".disclosure-triangle").click();
-                            expect($(".disclosure-triangle").hasClass("expanded")).toBeFalsy();
-                            // Check that all results are hidden
-                            expect($(".bottom-panel-table tr[data-file-index=0][data-match-index]:hidden").length).toEqual(7);
-                            expect($(".bottom-panel-table tr[data-file-index=1][data-match-index]:hidden").length).toEqual(4);
-                            $(".disclosure-triangle").click();
-                            expect($(".disclosure-triangle").hasClass("expanded")).toBeTruthy();
-                            expect($(".bottom-panel-table tr[data-file-index=0][data-match-index]:visible").length).toEqual(7);
-                            expect($(".bottom-panel-table tr[data-file-index=1][data-match-index]:visible").length).toEqual(4);
-                        });
-                    });
                 });
             });
         });
